Setting up typings
