+++ tsconfig.json
@@ -16,7 +16,7 @@
     "noUnusedLocals": true
   },
   "filesGlob": [
-    "typings/index.d.ts",
+    "typings/*.d.ts",
     "src/**/*.ts",
     "src/**/*.tsx"
   ],
