+++ tslint.json
@@ -18,11 +18,11 @@
     "no-inferrable-types": [ true ],
     "no-internal-module": true,
     "no-magic-numbers": [ true, -1, 0, 1, 2 ],
-    "no-namespace": true,
+    "no-namespace": false,
     "no-non-null-assertion": true,
     "no-reference": true,
     "no-unnecessary-type-assertion": true,
-    "no-var-requires": true,
+    "no-var-requires": false,
     "only-arrow-functions": [ true, "allow-declarations" ],
     "prefer-for-of": true,
     "promise-function-async": true,
